backend/requirements.txt
fastapi==0.104.1
uvicorn==0.24.0
sqlalchemy>=2.0.31,<3.0.0
psycopg2-binary==2.9.9
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
python-multipart==0.0.6
python-dotenv==1.0.0
httpx==0.25.2
pydantic>=2.7.3,<3.0.0
pydantic-settings==2.1.0
asyncpg==0.29.0
email-validator==2.1.0.post1
openai>=1.33.0,<2.0.0
tiktoken==0.5.2
